requirements.txt
Flask==2.0.3
flask-cors==3.0.10
numpy==1.21.4
Pillow==8.4.0
tensorflow==2.6.0
opencv-python==4.5.3.20210927
